added SweetAlerts 2 to the delete and create saves functions

+++ server/public/scripts/controllers/user.controller.js
@@ -8,7 +8,37 @@ myApp.controller('UserController', function (UserService) {
   UserService.getOthers();
 
   vm.deleteBuild = function (id) {
-    UserService.deleteBuild(id)
+    swal({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      confirmButtonClass: 'btn btn-success',
+      cancelButtonClass: 'btn btn-danger',
+      buttonsStyling: true
+    }).then(function () {
+      swal(
+        'Delete Confirmed',
+        'Your build has been successfully deleted.',
+        'success',
+        UserService.deleteBuild(id)
+      )
+    }, function (dismiss) {
+      // dismiss can be 'cancel', 'overlay',
+      // 'close', and 'timer'
+      if (dismiss === 'cancel') {
+        swal(
+          'Delete Cancelled',
+          'Your build is still saved',
+          'error'
+        )
+      }
+    })
+    
   }
   vm.putBuild = function (info, build) {
     UserService.putBuild(info, build)
